Accept reports in JSON-format - closes #7
ACRA can send the reports in JSON-format, too. The json-report will be sent within the body of the http request instead of key-value-pairs as post parameters.

The ApiController now differs between the content type of the request. If the content type is json the request content will be evaluated otherwise it will continue as before.

The incoming json object will be converted to an array in the very same format of an incoming post request with key-value parameters so the application can handle it.