Added phinx as new dependency to make db migrations easily
Phinx is a PHP database migration tool (see: https://phinx.org/). It allows us to write a migrate an existing database scheme to a new code base.

To create a new migration run 'vendor/bin/phinx create <MigrationName>' this will create a new php-class in the directory db/migrations (this directory should be created first)

To migrate an existing database run the shell script './runMigration.sh'. This makes sure that the db configuration defined in the .env-file will be used.

The file phinx.yml provides the configuration for phinx. In the current configuration it expects that some environment variables are set. The script 'runMigration.sh' ensures this expectation by loading the .env-file and exporting the variables.