Feature: Hide reports
For a better overview it is really helpful to be able to mark some reports as hidden. Especially if you do not want to delete them directly.

Therefore I added two more actions (Show/Hide) to the reports view. This will send the selected ids to the reports or showReports action of the new HideController. The reports will be flagged in the database as hidden.

To flag the reports I added a new column 'hidden' to the database schema and provided a migration class to migrate an existing database with phinx (see previous commit or https://phinx.org/).

So that the new field makes sense I changed all select statements concerning the reports table to skip "hidden" rows. To display hidden reports if necessary you can click on 'Show hidden reports' in the reports view.